Generate breaks for measurements below the limit of quantification

Breaks that are < lloq are removed. If the lowest break is removed if it is too close to the lloq.

Usage

breaks_blq_general(lloq, breakfun, trans = identity, ...)

Arguments

lloq

The value of the lower limit of quantification as a numeric scalar

breakfun

The function used for normal scale breaks if the lloq were not present.

trans

A parameter translation function (typically either identity for linear scale or log for log scale).

...

passed as breakfun(n=n, ...)

Value

A function for calculating breaks with arguments x and n

Details

For ggplot2 scales. This is not usually used directly. See blq_trans() and blq_log10_trans() for the functions that are more commonly used.

See also

Examples

breaks_blq_general(lloq=3, breakfun=scales::breaks_extended)(1:100, n=5)
#> [1]   3  25  50  75 100