pmxTools: Pharmacometric and Pharmacokinetic Toolkit

Pharmacometric tools for common data analytical tasks; closed-form solutions for calculating concentrations at given times after dosing based on compartmental PK models (1-compartment, 2-compartment and 3-compartment, covering infusions, zero- and first-order absorption, and lag times, after single doses and at steady state, per Bertrand & Mentre (2008) https://www.facm.ucl.ac.be/cooperation/Vietnam/WBI-Vietnam-October-2011/Modelling/Monolix32_PKPD_library.pdf); parametric simulation from NONMEM-generated parameter estimates and other output; and parsing, tabulating and plotting results generated by Perl-speaks-NONMEM (PsN).

See also

Author

Maintainer: Justin Wilkins justin.wilkins@occams.com (ORCID)

Authors:

Bill Denney wdenney@humanpredictions.com (ORCID)

Rik Schoemaker rik.schoemaker@occams.com

Other contributors:

Satyaprakash Nayak sn248@cornell.edu [contributor]

Leonid Gibiansky lgibiansky@quantpharm.com [contributor]

Andrew Hooker Andrew.Hooker@farmbio.uu.se [contributor]

E. Niclas Jonsson niclas.jonsson@pharmetheus.com [contributor]

Mats O. Karlsson mats.karlsson@farmbio.uu.se [contributor]

John Johnson johndjohnson@gmail.com [contributor]