Read (single or) multiple NONMEM tables from a single file

Read (single or) multiple NONMEM tables from a single file

read_nm_multi_table(
  fileName,
  header = TRUE,
  ...,
  simplify = TRUE,
  table_start_pattern = "^TABLE NO"
)

Arguments

fileName
The filename to read from

header, ...
Arguments passed to read.table

simplify
If a single table is present, return a data.frame instead of a list of data.frames?

table_start_pattern
What should be found to start a new table?

Value

A list of data.frames, or if only one is present and simplify=TRUE, a data.frame.

See also

Examples

if (FALSE) {
read_nm_multi_table("run1.cov", row.names=1)
}