Read PsN SCM output into a format suitable for further use.

read_scm returns a summary of a Perl-speaks-NONMEM (PsN, https://uupharmacometrics.github.io/PsN/) SCM (stepwise covariate modeling) procedure. It depends on the presence of scmlog.txt and short_scmlog.txt files in the specified directory.

Usage

read_scm(dir, startPhase = "forward")

Arguments

dir

A PsN SCM folder (containing scmlog.txt and short_scmlog.txt).

startPhase

Where to start collating the output; can be "forward" (the default) or "backward".

Value

A list of data frames, containing

forward

all models evaluated during the forward inclusion step of covariate model building

forwardSummary

the covariate relationships selected at each forward step

forwardP

the P-value used for inclusion during the forward inclusion step

backward

all models evaluated during the backward elimination step of covariate model building

backwardSummary

the covariate relationships eliminated at each backward step

backwardP

the P-value used for exclusion during the backward elimination step

See also

Author

Justin Wilkins, justin.wilkins@occams.com

Examples

if (FALSE) { # \dontrun{
scm <- read_scm("E:/DrugX/ModelDevelopment/scm310")
} # }